migrated end of run report logging functionality to ms_io
