+++ casanovo/denovo/model_runner.py
@@ -19,8 +19,6 @@ from lightning.pytorch.callbacks import ModelCheckpoint
 
 from ..config import Config
 from ..data import ms_io
-from ..data.logger_io import LogPredictionWriter
-from ..data.prediction_io import PredictionMultiWriter
 from ..denovo.dataloaders import DeNovoDataModule
 from ..denovo.model import Spec2Pep
 
@@ -148,22 +146,19 @@ class ModelRunner:
         -------
         self
         """
-        mztab_writer = ms_io.MztabWriter(Path(output).with_suffix(".mztab"))
-        mztab_writer.set_metadata(
+        self.writer = ms_io.MztabWriter(Path(output).with_suffix(".mztab"))
+        self.writer.set_metadata(
             self.config,
             model=str(self.model_filename),
             config_filename=self.config.file,
         )
 
-        log_writer = LogPredictionWriter(logger)
-        log_writer.log_start_time()
-        self.writer = PredictionMultiWriter([mztab_writer, log_writer])
         self.initialize_trainer(train=False)
         self.initialize_model(train=False)
         self.model.out_writer = self.writer
 
         test_index = self._get_index(peak_path, False, "")
-        mztab_writer.set_ms_run(test_index.ms_files)
+        self.writer.set_ms_run(test_index.ms_files)
         self.initialize_data_module(test_index=test_index)
         self.loaders.setup(stage="test", annotated=False)
         self.trainer.predict(self.model, self.loaders.test_dataloader())
